requirements.txt
Django==1.9.6
django-bootstrap-form==3.2.1
googlemaps==2.4.4
PyYAML==3.11
requests==2.9.1